backend/assets/assets/requirements.txt
asgiref==3.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.12
django-auto-prefetching==0.2.11
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-import-export==2.8.0
django-rest-multiple-models==2.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.1.0
PyJWT==2.3.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
